♻️ Milliseconds instead of seconds for MR.

+++ syncapi/storage/mrd/types.go
@@ -1,6 +1,6 @@
 package mrd
 
 type StateEvent struct {
-	Hidden   bool `json:"hidden"`
-	ExpireTs int  `json:"expire_ts"`
+	Hidden   bool  `json:"hidden"`
+	ExpireTs int64 `json:"expire_ts"`
 }
